requirements.txt
numpy==1.24.3
tflite-runtime==2.14.0
flask==3.0.2
pillow==11.2.1
requests==2.32.3
